% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{defs.svg}
\alias{defs.svg}
\title{make svg defs}
\usage{
defs.svg(content = NULL)
}
\arguments{
\item{content}{a character or a list, group content}
}
\value{
the characher type of svg element
}
\description{
make svg defs
}
\examples{
defs.svg(content = "this is a svg element")
content <- list(svg1 = "this is a svg element",
                svg2 = "this is a svg element")
defs.svg(content = content)

}
