% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group.svg}
\alias{group.svg}
\title{make svg group}
\usage{
group.svg(id = NULL, content = NULL)
}
\arguments{
\item{id}{a character, group id}

\item{content}{a character or a list, group content}
}
\value{
the characher type of svg element
}
\description{
make svg group
}
\examples{
group.svg(id = "group_1", content = "this is a svg element")
content <- list(svg1 = "this is a svg element",
                svg2 = "this is a svg element")
group.svg(id = "group_1", content = content)

}
