% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{lim.axis.svg}
\alias{lim.axis.svg}
\title{Generate SVG element of axis}
\usage{
lim.axis.svg(x = NULL, stroke = "#000000", stroke.width = 1,
  line.length = 100, axis.font.size = 8, id = NULL, unit = NULL)
}
\arguments{
\item{x}{a vector, the range of your number}

\item{stroke}{a number, the line stroke of the axis}

\item{stroke.width}{a number, the line stroke of the axis}

\item{axis.font.size}{a number, the axis font size of axis}

\item{length}{a number, the line length of the axis}
}
\value{
the characher type of svg element
}
\description{
This function will generate a axis form SVG element.
}
\examples{
lim.axis.1 <- lim.axis.svg(x = c(100, 900), id = "test")
pack.svg(pack.content = lim.axis.1, output.svg.name = "inst/test.svg")

lim.axis.2 <- lim.axis.svg(x = c(0, 20000), id = "test", unit = 4000, axis.font.size = 4)
pack.svg(pack.content = lim.axis.2, output.svg.name = "inst/test.svg")

}
