% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pack.svg}
\alias{pack.svg}
\title{pack svg}
\usage{
pack.svg(width = 1200, height = 800, output = F, output.name = NULL,
  content = content)
}
\arguments{
\item{width}{a number, width of the plot}

\item{height}{a number, height of the plot}

\item{content}{a character or a list, group content}
}
\value{
the characher type of svg element
}
\description{
pack svg
}
\examples{
pack.svg(content = "<text x=\\"10\\" y=\\"20\\"> this is a svg element </text>")
content <- list(svg1 = "<text x=\\"10\\" y=\\"20\\"> this is a svg element </text>",
                svg2 = "<text x=\\"10\\" y=\\"20\\"> this is a svg element </text>")
pack.svg(content = content)
pack.svg(content = "<text x=\\"10\\" y=\\"20\\"> this is a svg element </text>", output = T, output.name = "test.svg")

}
