% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text.svg}
\alias{text.svg}
\title{Generate polyline SVG element}
\usage{
\method{text}{svg}(x = NULL, y = NULL, content = "", fill = "#000000",
  stroke = "none", stroke.width = 1, font.family = "Arial",
  font.size = 8, font.weight = "normal", font.style = "normal",
  text.decoration = "none", word.spacing = "normal",
  letter.spacing = "normal", text.anchor = "start", rotate = NULL,
  text.path = NULL)
}
\arguments{
\item{x}{a number, x coordinate information}

\item{y}{a number, y corrdinate information}

\item{fill}{a character, color of the text, eg. "#000000"(default), "red"}

\item{stroke}{a characher, color of the rect text, eg. "#000000"(default), "red"}

\item{stroke.width}{a number, stroke width of the rect text, default: 1}

\item{font.family}{a character, font family of text, default: "Arial"}

\item{font.size}{a number, font size of text, default: 8}

\item{font.weight}{a character, font weight of text, eg. "normal"(default), "bold"}

\item{font.style}{a character, font style of text, eg. "normal"(default), "italic"}

\item{text.decoration}{a character, text decoration, eg. "none"(default), "underline", "overline", "line-through"}

\item{word.spacing}{a number or character, default: "normal"}

\item{letter.spacing}{a number or character, defailt: "normal"}

\item{text.anchor}{a character, eg. "start"(default), "middle", "end"}

\item{rotate}{a number, rotation angle of text}

\item{text.path}{a character, fit text path}

\item{ry}{a number, y coordinate of rounded rectangle}
}
\value{
the characher type of svg element
}
\description{
Generate polyline SVG element
}
\examples{
text.svg(x = 10, y = 20, content = "Hello Word", fill = "blue")
text.svg(x = 10, y = 20, content = "Hello Word", fill = "blue", rotate = 90, font.family = "Helvetica")

}
