% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{use.svg}
\alias{use.svg}
\title{use svg}
\usage{
use.svg(id = NULL, x = NULL, y = NULL, scale, rotate, translate, skewX,
  skewY, style.sheet = NULL, transform.sheet = NULL)
}
\arguments{
\item{id}{a character, target of the link}

\item{x}{a number, x transform coordinate}

\item{y}{a number, y transform coordinate}

\item{scale}{a number. transform scale of the object}

\item{rotate}{a vector, rotation of the object}

\item{translate}{a vector, translate of the object}

\item{skewX}{a number}

\item{skewY}{a number}

\item{style.sheet}{a vector or a chatacter, other style of the link, eg. "stroke-linecap: round"}

\item{transform.sheet}{a vector or a chatacter, other transform of the link,}
}
\value{
the characher type of svg element
}
\description{
The <use> element takes nodes from within the SVG document, and duplicates them somewhere else.
}
\examples{
use.svg(id = "target", x = 100, y = 200)
use.svg(id = "target", x = 100, y = 200, rotate = c(90, 100, 200))

}
