% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{defs.svg}
\alias{defs.svg}
\title{make svg defs}
\usage{
defs.svg(defs.content = NULL)
}
\arguments{
\item{defs.content}{a character or a list, group content}
}
\value{
the characher type of svg element
}
\description{
make svg defs
}
\examples{
defs.svg(defs.content = "<text x=\\"10\\" y=\\"20\\"> a svg element </text>")
defs.content <- list(svg1 = "<text x=\\"10\\" y=\\"30\\"> a svg element </text>",
                     svg2 = "<text x=\\"10\\" y=\\"40\\"> a svg element </text>")
defs.svg(defs.content = defs.content)

}
