% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{defs.svg}
\alias{defs.svg}
\title{make SVG defs element}
\usage{
defs.svg(defs.content = NULL)
}
\arguments{
\item{defs.content}{a character or a list, group content}
}
\value{
the characher type of SVG element
}
\description{
make SVG defs element
}
\examples{
defs.svg(defs.content = "<text x=\\"10\\" y=\\"20\\"> an SVG element </text>")
defs.content <- list(svg1 = "<text x=\\"10\\" y=\\"30\\"> an SVG element </text>",
                     svg2 = "<text x=\\"10\\" y=\\"40\\"> an SVG element </text>")
defs.svg(defs.content = defs.content)

}
