% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group.svg}
\alias{group.svg}
\title{make svg group}
\usage{
group.svg(id = NULL, group.content = NULL, style.sheet = NULL)
}
\arguments{
\item{id}{a character, group id}

\item{group.content}{a character or a list or a vector, group content}

\item{style.sheet}{a vector or a chatacter, other style of the group, eg. "stroke-linecap: round"}
}
\value{
the characher type of svg element
}
\description{
make svg group
}
\examples{
group.svg(id = "group_1", group.content = "this is a svg element")
group.content <- list(svg1 = "this is a svg element",
                      svg2 = "this is a svg element")
group.svg(id = "group_1", group.content = group.content)

}
